Add --scale-available option to scale up/to commands
By default if a cluster has only 5 available workers
(eg, 2 online / 3 offline) and is asked to scale to 6
the command will fail with a message complaining about
the cluster not having enough workers. The --scale-available
option tells mo-scaler to go ahead and scale up to the
available number of workers.